Implemented payment

+++ app/src/main/java/com/example/cginvoice/data/source/local/entitiy/invoice/PaymentEntity.kt
@@ -33,7 +33,7 @@ data class PaymentEntity(
             paymentId = paymentId.toInt(),
             paymentObjectId = paymentObjectId,
             invoiceId = invoiceId,
-            paymentDate = paymentDate,
+            paymentDate = paymentDate.toLong(),
             amount = amount,
             paymentMethod = paymentMethod,
             note = note
